Develop

+++ skynet-src/socket/socket_server_def.h
@@ -33,14 +33,14 @@ namespace skynet {
 // socket event type
 enum socket_event
 {
-    SOCKET_EVENT_DATA = 0,              // socket 正常数据
-    SOCKET_EVENT_CLOSE = 1,             // socket close 数据
-    SOCKET_EVENT_OPEN = 2,              // socket open 数据 (connect)
-    SOCKET_EVENT_ACCEPT = 3,            // socket accept 数据
-    SOCKET_EVENT_ERROR = 4,             // socket error 数据
-    SOCKET_EVENT_EXIT = 5,              // socket exit 数据
-    SOCKET_EVENT_UDP = 6,               // socket udp 数据
-    SOCKET_EVENT_WARNING = 7,           // socket warning 数据
+    SOCKET_EVENT_DATA = 0,              // socket data event
+    SOCKET_EVENT_CLOSE = 1,             // socket close event
+    SOCKET_EVENT_OPEN = 2,              // socket connect event
+    SOCKET_EVENT_ACCEPT = 3,            // socket accept event
+    SOCKET_EVENT_ERROR = 4,             // socket error event
+    SOCKET_EVENT_EXIT = 5,              // socket exit event
+    SOCKET_EVENT_UDP = 6,               // socket udp event
+    SOCKET_EVENT_WARNING = 7,           // socket warning event
     SOCKET_EVENT_RST = 8,               // only for internal use
 };
 
@@ -56,7 +56,7 @@ struct socket_object_interface
 struct socket_message
 {
     int socket_id = 0;                  //
-    uint64_t svc_handle = 0;            // skynet service handle, where message sendding, TODO: change to uint32_t
+    uint32_t svc_handle = 0;            // skynet service handle, where message sendding, TODO: change to uint32_t
     int ud = 0;                         // for accept: ud is new connection's fd; for data: ud is the size of data
     char* data_ptr = nullptr;           // data ptr
 };
